Remove FXIOS-15164 legacy first-run onboarding (intro + ToS)
Remove the legacy UIKit first-run onboarding path now that modern
onboarding is the default. Collapses the modernOnboardingUI branches in
the launch flow to modern-only; flag-off now falls through to the modern
variant instead of legacy.

- Delete IntroViewController, IntroViewModel, TermsOfServiceViewController
  and the OnboardingViewControllerState Redux wiring
- Collapse the launch branches in LaunchCoordinator, SceneCoordinator and
  LaunchType; remove the now-dead presentTermsOfService/presentIntroOnboarding
- Move the OnboardingOptions OptionSet into ActivityEventHelper and the
  Links URLs into TermsOfServiceManager (both were owned by deleted files)
- Update/remove the corresponding unit tests

The "What's New" upgrade flow still uses the shared card engine and is
retired in a follow-up.
