+++ firefox-ios/firefox-ios-tests/Tests/ClientTests/Helpers/IntroScreenManagerTests.swift
@@ -86,10 +86,10 @@ final class IntroScreenManagerTests: XCTestCase {
 
     // MARK: - onboardingVariant Tests
 
-    func testOnboardingVariant_whenBothFlagsDisabled_returnsLegacy() {
+    func testOnboardingVariant_whenBothFlagsDisabled_returnsModern() {
         setupNimbusFeatureFlags(enableModernUi: false, shouldUseJapanConfiguration: false)
         let subject = IntroScreenManager(prefs: prefs)
-        XCTAssertEqual(subject.onboardingVariant, .legacy)
+        XCTAssertEqual(subject.onboardingVariant, .modern)
     }
 
     func testOnboardingVariant_whenModernEnabledButJapanDisabled_returnsModern() {
@@ -104,11 +104,11 @@ final class IntroScreenManagerTests: XCTestCase {
         XCTAssertEqual(subject.onboardingVariant, .japan)
     }
 
-    func testOnboardingVariant_whenModernDisabledButJapanEnabled_returnsLegacy() {
-        // Japan configuration requires modern UI to be enabled
+    func testOnboardingVariant_whenModernDisabledButJapanEnabled_returnsModern() {
+        // Japan configuration requires modern UI to be enabled; otherwise it falls through to modern
         setupNimbusFeatureFlags(enableModernUi: false, shouldUseJapanConfiguration: true)
         let subject = IntroScreenManager(prefs: prefs)
-        XCTAssertEqual(subject.onboardingVariant, .legacy)
+        XCTAssertEqual(subject.onboardingVariant, .modern)
     }
 
     // MARK: - shouldShowVideoIntro Tests
